+++ src/termout.c
@@ -5295,8 +5295,10 @@ term_do_write(const char *buf, uint len, bool fix_status)
           term.curs.attr.attr &= ~FONTFAM_MASK;
           term.curs.attr.attr |= (cattrflags)gcode << ATTR_FONTFAM_SHIFT;
         }
+#define draw_powerline_geometric_symbols
 #ifdef draw_powerline_geometric_symbols
-#warning graphical results of this approach are unpleasant; not enabled
+// don't remember why this was not enabled...
+//#warning graphical results of this approach are unpleasant; not enabled
         else if (wc >= 0xE0B0 && wc <= 0xE0BF && wc != 0xE0B5 && wc != 0xE0B7) {
           // draw geometric full-cell Powerline symbols,
           // to avoid artefacts at their borders (#943)